Fix the loading of terminal commands.
Even though these are generally treated like individual "apps", they're loaded as a single component by the ExecutionEngine.
The definitive list of commands are defined by Util as this list is required in lots of different places. The loaders in the TerminalCommands component cross check its own definitions against this definitive list.