Terminal's fake input text field was causing scroll issues when the terminal window gains focus.
Because the fake input was in the UI tree at a specific point, it would scroll to that point. By moving the fake input outside of the scrollable div, giving it focus has no affect on the scroll offset. Move this logic to onFocusing instead of onFocus, which has no effect on anything, but I added the handler to the window options in the shell and it seemed generally useful to add.